refactor(plugin): relocate discovery + registry adapters from nebula-sandbox
Host-registry population belongs with the registry. `discovery`,
`plugin_toml`, `RemoteAction`/`RemoteActionFactory`, `DiscoveredPlugin`,
and `ProcessSandboxHandler` move from `nebula-sandbox` into
`nebula-plugin`, fixing the live canon violation where discovery mutated
`nebula_plugin::PluginRegistry` from outside the registry crate.

Layering: `nebula-sandbox` becomes the Plugin-Proto leaf transport (no
Business dependency). `nebula-plugin` (Business) -> `nebula-sandbox` /
`nebula-plugin-sdk` are legal downward edges; the `deny.toml [bans]`
wrappers are updated to admit them (folded here so `cargo deny check`
stays green at the commit the new edges first appear). The
`nebula-plugin` path dep is dropped from `crates/sandbox/Cargo.toml` to
break the dependency cycle the new edge would otherwise form (remaining
now-unused Business deps are slimmed in a follow-up commit).

`sandbox_error_to_action_error` consolidates into
`nebula-plugin::sandbox_bridge` as the single `SandboxError` ->
`ActionError` seam shared by `ProcessSandboxHandler` and the engine
`SandboxRunner` adapter (which consumes it via the legal engine ->
plugin downward edge); the engine-local copy is removed. The two
discovery integration tests and the classification guards move with
their code — no test dropped, total count unchanged.
