+++ crates/sandbox/Cargo.toml
@@ -17,7 +17,6 @@ nebula-action = { path = "../action" }
 nebula-core = { path = "../core" }
 nebula-credential = { path = "../credential" }
 nebula-metadata = { path = "../metadata" }
-nebula-plugin = { path = "../plugin" }
 nebula-plugin-sdk = { path = "../plugin-sdk" }
 nebula-resource = { path = "../resource" }
 nebula-schema = { path = "../schema" }
